+++ packages/components/src/components.ts
@@ -9,8 +9,15 @@ import HdsAlert from './components/hds/alert/index.ts';
 import HdsAppFooter from './components/hds/app-footer/index.ts';
 import HdsAppFrame from './components/hds/app-frame/index.ts';
 import HdsApplicationState from './components/hds/application-state/index.ts';
+import HdsApplicationStateBody from './components/hds/application-state/body.ts';
+import HdsApplicationStateFooter from './components/hds/application-state/footer.ts';
+import HdsApplicationStateHeader from './components/hds/application-state/header.ts';
+import HdsApplicationStateMedia from './components/hds/application-state/media.ts';
 import HdsBadge from './components/hds/badge/index.ts';
 import HdsBadgeCount from './components/hds/badge-count/index.ts';
+import HdsBreadcrumb from './components/hds/breadcrumb/index.ts';
+import HdsBreadcrumbItem from './components/hds/breadcrumb/item.ts';
+import HdsBreadcrumbTruncation from './components/hds/breadcrumb/truncation.ts';
 import HdsButton from './components/hds/button/index.ts';
 import HdsButtonSet from './components/hds/button-set/index.ts';
 import HdsCard from './components/hds/card/container.ts';
@@ -99,8 +106,15 @@ export {
   HdsAppFooter,
   HdsAppFrame,
   HdsApplicationState,
+  HdsApplicationStateBody,
+  HdsApplicationStateFooter,
+  HdsApplicationStateHeader,
+  HdsApplicationStateMedia,
   HdsBadge,
   HdsBadgeCount,
+  HdsBreadcrumb,
+  HdsBreadcrumbItem,
+  HdsBreadcrumbTruncation,
   HdsButton,
   HdsButtonSet,
   HdsCard,