+++ packages/components/src/components/hds/application-state/footer.ts
@@ -3,9 +3,11 @@
  * SPDX-License-Identifier: MPL-2.0
  */
 
-import Component from '@glimmer/component';
+import TemplateOnlyComponent from '@ember/component/template-only';
 import type { ComponentLike } from '@glint/template';
 import type { HdsLinkStandaloneSignature } from '../link/standalone';
+import type { HdsButtonSignature } from '../button';
+import type { HdsDropdownSignature } from 'src/components/hds/dropdown';
 
 export interface HdsApplicationStateFooterSignature {
   Args: {
@@ -14,6 +16,8 @@ export interface HdsApplicationStateFooterSignature {
   Blocks: {
     default?: [
       {
+        Button?: ComponentLike<HdsButtonSignature>;
+        Dropdown?: ComponentLike<HdsDropdownSignature>;
         LinkStandalone?: ComponentLike<HdsLinkStandaloneSignature>;
       },
     ];
@@ -21,31 +25,7 @@ export interface HdsApplicationStateFooterSignature {
   Element: HTMLDivElement;
 }
 
-export default class HdsApplicationStateFooterComponent extends Component<HdsApplicationStateFooterSignature> {
-  /**
-   * Indicate if the footer should have a top border or not.
-   *
-   * @param hasDivider
-   * @type {boolean}
-   * @default false
-   */
-  get hasDivider(): boolean {
-    return this.args.hasDivider ?? false;
-  }
+const HdsApplicationStateFooterComponent =
+  TemplateOnlyComponent<HdsApplicationStateFooterSignature>();
 
-  /**
-   * Get the class names to apply to the component.
-   * @method classNames
-   * @return {string} The "class" attribute to apply to the component.
-   */
-  get classNames(): string {
-    const classes = ['hds-application-state__footer'];
-
-    // add a class based on the existence of @hasDivider argument
-    if (this.hasDivider) {
-      classes.push(`hds-application-state__footer--has-divider`);
-    }
-
-    return classes.join(' ');
-  }
-}
+export default HdsApplicationStateFooterComponent;