+++ packages/components/src/components/hds/breadcrumb/index.ts
@@ -5,16 +5,28 @@
 
 import Component from '@glimmer/component';
 
+export interface HdsBreadcrumbSignature {
+  Args: {
+    ariaLabel?: string;
+    itemsCanWrap?: boolean;
+    didInsert?: () => void;
+  };
+  Blocks: {
+    default: [];
+  };
+  Element: HTMLElement;
+}
+
 const NOOP = () => {};
 
-export default class HdsBreadcrumbComponent extends Component {
+export default class HdsBreadcrumbComponent extends Component<HdsBreadcrumbSignature> {
   /**
    * @param onDidInsert
    * @type {function}
    * @default () => {}
    */
-  get didInsert() {
-    let { didInsert } = this.args;
+  get didInsert(): () => void {
+    const { didInsert } = this.args;
 
     if (typeof didInsert === 'function') {
       return didInsert;
@@ -28,7 +40,7 @@ export default class HdsBreadcrumbComponent extends Component {
    * @type {boolean}
    * @default true
    */
-  get itemsCanWrap() {
+  get itemsCanWrap(): boolean {
     return this.args.itemsCanWrap ?? true;
   }
 
@@ -37,7 +49,7 @@ export default class HdsBreadcrumbComponent extends Component {
    * @type {string}
    * @default 'breadcrumbs'
    */
-  get ariaLabel() {
+  get ariaLabel(): string {
     return this.args.ariaLabel ?? 'breadcrumbs';
   }
 
@@ -46,8 +58,8 @@ export default class HdsBreadcrumbComponent extends Component {
    * @method Breadcrumb#classNames
    * @return {string} The "class" attribute to apply to the component.
    */
-  get classNames() {
-    let classes = ['hds-breadcrumb'];
+  get classNames(): string {
+    const classes = ['hds-breadcrumb'];
 
     // add a class based on the @itemsCanWrap argument
     if (this.itemsCanWrap) {